requirements.txt
openai
requests
nicegui
python-docx
pypdf